feat(NavBarSheet): not render until isMobile defined; sheetCloseWrapper for mobile
